Undo this renaming of org_reader to org_emacs_reader.. API shouldn't change like this
I have a bunch of lisp haks myself, the python client will ignore all of those.
Also the native python one requires an extra dependency. It's much faster though
