--- org_emacs_reader/__init__.py
@@ -1,1 +1,0 @@
-from .org_emacs_reader import *
